+++ test/conftest.py
@@ -62,11 +62,14 @@ class VolumeFile:
     data_blob_id: str
     mode: int
 
+
 @dataclasses.dataclass
 class GrpcErrorAndCount:
-    """ Helper class that holds a gRPC error and the number of times it should be raised. """
+    """Helper class that holds a gRPC error and the number of times it should be raised."""
+
     grpc_error: Status
     count: int
+
 
 # TODO: Isolate all test config from the host
 @pytest.fixture(scope="function", autouse=True)
@@ -1043,6 +1046,10 @@ class MockClientServicer(api_grpc.ModalClientBase):
         if function_defn.schedule:
             self.function2schedule[function_id] = function_defn.schedule
 
+        warnings = []
+        if int(function_defn.experimental_options.get("warn_me", "0")):
+            warnings.append(api_pb2.Warning(message="You have been warned!"))
+
         await stream.send_message(
             api_pb2.FunctionCreateResponse(
                 function_id=function_id,
@@ -1069,6 +1076,7 @@ class MockClientServicer(api_grpc.ModalClientBase):
                     class_parameter_info=function_defn.class_parameter_info,
                     function_schema=function_defn.function_schema,
                 ),
+                server_warnings=warnings,
             )
         )
 
@@ -1285,8 +1293,10 @@ class MockClientServicer(api_grpc.ModalClientBase):
         # update function definition
         fn_definition = self.app_functions[req.function_id]
         assert isinstance(fn_definition, api_pb2.Function)
-        fn_definition.warm_pool_size = req.warm_pool_size_override  # hacky
-
+        # Hacky that we're modifying the function definition directly
+        # In the server we track autoscaler updates separately
+        fn_definition.warm_pool_size = req.warm_pool_size_override
+        fn_definition.autoscaler_settings.MergeFrom(req.settings)
         await stream.send_message(api_pb2.FunctionUpdateSchedulingParamsResponse())
 
     ### Image
@@ -1332,6 +1342,9 @@ class MockClientServicer(api_grpc.ModalClientBase):
     async def MountPutFile(self, stream):
         request: api_pb2.MountPutFileRequest = await stream.recv_message()
         if request.WhichOneof("data_oneof") is not None:
+            if request.data.startswith(b"large"):
+                # Useful for simulating a slow upload, e.g. to test our checks for mid-deploy modifications
+                await asyncio.sleep(2)
             self.files_sha2data[request.sha256_hex] = {"data": request.data, "data_blob_id": request.data_blob_id}
             self.n_mount_files += 1
             await stream.send_message(api_pb2.MountPutFileResponse(exists=True))
@@ -1478,7 +1491,7 @@ class MockClientServicer(api_grpc.ModalClientBase):
     async def SandboxCreate(self, stream):
         request: api_pb2.SandboxCreateRequest = await stream.recv_message()
         self.sandbox = await asyncio.subprocess.create_subprocess_exec(
-            *request.definition.entrypoint_args,
+            *(request.definition.entrypoint_args or ["sleep", f"{48 * 3600}"]),
             stdout=asyncio.subprocess.PIPE,
             stderr=asyncio.subprocess.PIPE,
             stdin=asyncio.subprocess.PIPE,